+++ src/main/scala/model/spawner/package.scala
@@ -23,23 +23,3 @@ package object spawner:
       GenericSpawner(3, spawnable = SpecificSpawners.PlaneStrategy(height))
     ):_*
   )
-
-  /**
-   * Extension methods for DeltaTime
-   */
-  extension(interval: DeltaTime)
-
-    /**
-     * The hyperbole function that given a timePassed return the updated missile
-     */
-    def ~(timePassed: DeltaTime) : DeltaTime =
-      interval - ((interval * (1 / ( 1 + Math.exp(((-1 * timePassed) / 10))))) - (interval / 2))
-    /**
-     * Function that given a condition and a world può inficiare positivamente sulle lezaione.
-     * @param condition The condition to which we can subscreibe
-     * @param mapper
-     * @return
-     */
-    def mapIf(condition: () => Boolean)(mapper: (DeltaTime) => DeltaTime): DeltaTime = condition match
-      case _ if condition() => mapper(interval)
-      case _ => interval
